+++ game-witcher3/CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.6.2] - 2024-xx-xx
+## [1.6.4] - 2024-06-12
+
+- Fixed incorrect detection of load order entry names for mixed mods
+
+## [1.6.3] - 2024-06-11
+
+- Fixed the installer for mixed mods not executing (been broken for a while)
+- Fixed mod installation for mods that provide both dlc and mods folders and do not contain menu mod information.
+
+## [1.6.2] - 2024-05-29
 
 - Fixed crash when attempting to create load order file, and user has insufficient permissions